Bring common code into lib.rs
